Restructure project to make use of packages
