--- src/main/java/IllegalArgumentsException.java
@@ -1,4 +1,0 @@
-package PACKAGE_NAME;
-
-public class IllegalArgumentsException {
-}
