+++ src/main/java/com/Elsria/CommandType.java
@@ -1,3 +1,5 @@
+package com.Elsria;
+
 import java.util.Arrays;
 import java.util.Set;
