+++ src/main/java/com/Elsria/Event.java
@@ -1,3 +1,5 @@
+package com.Elsria;
+
 public class Event extends Task {
     private String start;
     private String end;
