+++ src/main/java/com/Elsria/Task.java
@@ -1,3 +1,5 @@
+package com.Elsria;
+
 public abstract class Task {
     protected String task;
     protected boolean isDone;
